requirements.txt
Flask==3.1.2
Flask_Login==0.6.2
flask_sqlalchemy==3.1.1
numpy==2.3.5
opencv_contrib_python==4.12.0.88
Pillow==12.0.0
Requests==2.32.5
Werkzeug==3.1.4
